Rename files and fix exports
